Adding missing facts for imports (1/2)
Summary:
While having a look at the regression tests, I notice I am missing some import facts. In particular, for cases like
```
import a.b.c
```
The expected behaviour is to add imports and name facts for the three modules `a`, `a.b` and `a.b.c`.

In this diff, I am doing a refactoring to split the logic for `ImportStatement` and `ImportStatementFrom` which I am doing to better implement the fix which only apply for the former node type.

Reviewed By: aahanaggarwal

Differential Revision: D80404090

fbshipit-source-id: bea77b75742aa2cc5878d9cb3093fb5022c83202