Use `Self` in `Attribute` impl
Summary:
I am considering a significant refactor to split Attribute into
a simple enum for `attr.rs`, whose responsibility is handling
non-class-level attribute logic, versus the class attribute logic
that logically belongs in `class_field` (or possibly some other module
in the `alt/class` code).

My thinking is that class attribute lookup is very complex because of
descriptors, properties, qualifiers, and so on. General attribute is
also complex because of unions, module lookup fallbacks, getattr fallback
(which isn't restricted to classes, it has the same semantics for
modules and hence belongs in the general layer rather than the class-specific
layer).

Splitting them up will leave us with two complex but cleanly separated layers;
the current code is hard to work with, in my opinion, in part because the two
layers are not well separated and in particular several of the really difficult
cases involve throwing data over the fence between `class_field` logic and
`attr` logic by constructing complicated data structures that make the descriptor
handling in particular a little brittle.

I'm still not 100% sure about the refactor, but using `Self` in the existing
impl makes it easier and has no downside, so making that tweak now.

Reviewed By: yangdanny97

Differential Revision: D80465936

fbshipit-source-id: 499aa931d76e9f783e980474e6f383907e5df279