Quantified call target
Summary:
If we have a quantified type var and try to use it as a function, we need to:
1) properly break up the bounds or constraints into a union of CallTargets before we look up the `__call__` method
2) substitute Self w/ the type var, instead of with the concrete bound class

for step 2, we don't substitute with the original bound (which could be a union). instead, we synthesize a bound

```
class A(Protocol):
    def __call__(self) -> Self: ...
class B(Protocol):
    def __call__(self) -> Self: ...
T = TypeVar('T', bound=A | B)
def func(a: T) -> T:
    return a()
```

In this example, the Self in A gets replaced with `TypeVar(T, bound=A)` and the Self in B gets replaced with `TypeVar(T, bound=B)`

Both of them are subtypes of `TypeVar(T, bound=A|B)` due to the is_subset_eq relation implemented in the previous diff, so the return type is satisfied.

Reviewed By: stroxler

Differential Revision: D80535459

fbshipit-source-id: a63da8c6244b89c480ec219df6cc04e76d3703f1