Do not completely give up if one pattern's file listing failed
Summary:
Currently, when we have multiple pattern includes (very common if we have `project_includes = ["**/*.py", "**/*.pyi"]`), it will fail if anyone of them fail. This is quite extreme given that it can fail if one of them simply doesn't return any result. This is likely unintended, and likely just a result of lazily adding `?` to deal with results.

This diff specifically factor out `files_allow_empty_match`, so that it can be used for `Globs` to not fail all of the matching when one of the glob has empty match. I kept the `files` function the same.

Reviewed By: connernilsen, samwgoldman

Differential Revision: D80294842

fbshipit-source-id: be6fd04c613d6a9de0aa9a0b74f1084b20d32ffa