Insert `BindingExpect::TypeCheckExpr` for all valid subscript exprs in base class list
Summary:
hammer_cat

In D80387573, I took away the bound check when instantiating a generic class type within the base class list. This diff is one attempt to bring it back.

Basically, the way it works is that when inspecting `BaseClassExpr`s in class base list calculation, I've 100% suppressed all type errors. Instead, I emit separate expect binding for all valid `BaseClassExpr`s during the binding phase, and have that binding be responsible in actually performing user-facing type checking on the those exprs. Since normal type checking always validate bound when instantiating generic classes, we'll make sure we get the validation in these expect bindings.

Obviously, the reason why I don't like the approach is that it uses a big hammer to address a small issue. In theory, what we need here is just to "recover" the bound validation logic which is only a small part of the base expr type checking logic. What we end up doing here though is to duplicate *all* base expr type checking logic. But what makes it hard to find an alternative is that whether the bound validation logic is going to happen or not can't be reliably determined syntactically at the binding phase. We truly need type-level info which means that the recovery has to be done in answers phase, but at the same time I'm also not aware of a good way to create "delayed" computation on-the-fly (which only included the bound validation part of the logic) while we are in the process of solving the class base binding.

If anyone has alternative ideas to implement the recovery please do let me know.

Reviewed By: stroxler

Differential Revision: D80487859

fbshipit-source-id: 83d271bbc5f7b29d79175acf6d2b74e520e8f7b8