Add validate_by alias testcases
Summary:
We have to collect the validate_by_name and validate_by_alias keywords from the class.

The behavior there will affect the downstream behavior of the init function.

My thought is to add that info in the pydantic metadata binding as a first step. However, we must also know about the field name and be able to pass it to the init function. I will look into what we can do here so that the downstream code knows to add the field name to the initializer code so that we do not get a missing keyword error.

Reviewed By: rchen152

Differential Revision: D80570001

fbshipit-source-id: dcb574473a64394c976c361a73b37bdbd539cc74