Move `resolve_named_tuple_element`
Summary:
**About the next few diffs**

After staring at the attribute code in detail for a while, I've realised one of
the problems we have in Pyrefly is that we're using it for some things where it
is the wrong layer of abstraction.

The role of `attr.rs`, in my opinion, is essentially to model the full complexity
of attribute lookups which includes:
- unions (and we want to substantively change this!)
- special type system entities like typevars and modules
- special language features like getattr fallback

But we are using it in a few places to look up special kinds of *class* attributes,
manually creating the attribute base of an instance and then threading through all
of the irrelevant special-case logic.

One of the things we need - especially before we can make the attribute code union-native
in general - is to stop doing this... in many cases we know aprior that the base is *not*
a union or any other special case, and the caller needs to depend on this. That means
we need to work at the appropriate, lower level of abstraction of class field resolution.

**About this diff**

One of the callsites I think doesn't belong at `attr` level is named tuple element
resolution. We're always looking up a class field on an instance here, we don't want
all the other stuff in the way.

Reviewed By: yangdanny97

Differential Revision: D80572535

fbshipit-source-id: c330841bb940a5ca6e4104e8e6a7402f74471233