Move `type_of_enum_value` to the class field code
Summary:
**About this stack**

See previous diff for details; I'm working on layering class field and attr logic
more clearly and avoiding special-case calls that don't really want the general
attribute logic - those ought to directly operate on classes rather than going through
all the attribute plumbing.

**About this diff**

`type_of_enum_value` is another of these cases that shouldn't be in `attr.rs`.

In addition to moving it (and making it private, as a bonus!) I added a todo because
the doc comment discusses semantics that aren't currently enforced, but if we look at
how this gets used (the only callsite is inside of class field analysis, checking
enum members *against* the `_value_`) it isn't really clear that this is a problem.

Reviewed By: yangdanny97

Differential Revision: D80572536

fbshipit-source-id: 132085f55597772b6901bcef140d10f330339c2e