Collect validate_by data
Summary:
We add the info in Pydantic metadata for now since that gets passed to the dataclass code. However, I think that we can follow up with a refactor that does not propagate the entire structure to class metadata, since we're not using everything there.

We may need an intermediate structure for storing the things we need in the metadata level and then another for the answers level.

Secondly, we will go back to the place where we handled validation_aliases and probably generalize that logic to handle validate_by_alias/name logic and extend the logic downstream to handle it.

Reviewed By: rchen152

Differential Revision: D80581837

fbshipit-source-id: 60bd295449a0463e41ebd4b76eb86b52ffacda82