Move `do_check` into a better-named method
Summary:
Using nested functions makes the code harder to read because
we're passing around things that only have to be passed because
this isn't a method that can access `self` directly.

In addition, I'm going to need more power here in order to fix
nested class handling, where raw use of qnames doesn't work
properly.

Extracting the helper makes everything more readable and the
refactors we want to perform easier.

Reviewed By: vladima

Differential Revision: D80621409

fbshipit-source-id: ea3079e8466f37d280d6bd24dc3ec8190365da5a