Allow `is_subtype` to handle nested classes (mostly)
Summary:
In Python, qualified names are not actually unique identifiers, because
it's possible for attributes in `__init__` to shadow submodules so that
two separate values have the same qualified name, and importing that name
differently gives you different answers.

We cannot and will not handle this in our current query implementation, which
is intended for backward compatibility with Pyre1 - we basically only support
the subset of Python that doesn't have collisions.

This diff makes `is_subset` in the query API robust to nested classes, assuming
the codebase has no collisions: we can resolve a qualified name in that case by
just recursively backing off its prefix looking for the longest prefix that is
a valid module name.

Reviewed By: vladima

Differential Revision: D80621402

fbshipit-source-id: c89a2e2e1d29d59291b49fccb41471e00b3f6ed8