fix go-to definition on .py imports when not loaded in state
Summary:
## Background:

D80145015 started supporting go-to definition for modules that are not in tdeps: namely, files from stdlib `.py` functions are not analyzed for typechecking, instead only their `.pyi` counterpart is analyzed.

My implementation assumed equal or more go-to definitions working since we always fall back to .pyi when we can't find a .py.

## Problem:

Go-to definition works on some of these, but not all of these. It does not work when the handle is not yet loaded in the state, as `get_module_info` assumes handles are already loaded (it eventually calls [get_module_inner](https://fburl.com/code/srvst3fs)).

In this case we do find a .py, we just can't open the handle, so we end up with less total go-to definition (bad).

## Solution:

Call a function that will do something to load this file in the state. In this diff I picked `get_exports`

Downsides?

I think there are none since get_exports is free if we've already analyzed this file. If we haven't, we have to load it in state somehow.

todo:
Figure out how to test this + rethink some APIs to make this issue never happen again.

Reviewed By: yangdanny97

Differential Revision: D80623516

fbshipit-source-id: b7644b56afd8a814328e38fd78caaf9d999b4fc1