Don't allow adding extra items when inheriting from a closed TypedDict
Summary:
For https://github.com/facebook/pyrefly/issues/946.

Note that it looks like this change introduces a conformance regression, but it's actually a bug in the conformance tests: there's a case that's supposed to exercise inheriting from a TypedDict with closed=True and setting closed=False, and it conflates an error from changing the closed-ness with an error from extending the closed parent with an extra item (because pyright reports the latter error on the `class ...` line). I'll fix this upstream, but I plan to work my way through the rest of the extra items tests first, in case I find additional issues.

Reviewed By: yangdanny97

Differential Revision: D80577490

fbshipit-source-id: adfb0eed1793e496ce675b24a9cbaca9149409e2