Kill `resolve_as_instance_method_with_attribute_inner`
Summary:
This had a pretty hard-to-read name and added to the number of hoops you
had to jump through to understand the code.

I also realized it's only there because we were decomposing an attribute that we
already had, and we added a helper function to make the logic work with the
decomposition in place, but we can just simplify and not decompose instead.

This is very useful because if we move class-field specific information out of
`attr.rs`, then the type is going to change and having multiple helpers that
couple tightly to the exact data structure makes refactoring hard.

Reviewed By: grievejia

Differential Revision: D80623182

fbshipit-source-id: 9095278cb835243579cc886b9d27a8369b123460