Add a helper function `resolve_instance_method`
Summary:
The existing `resolve_as_instance_method` is a wierd implementation
detail of attribute lookup, we don't want to be relying on it from
disparate parts of the codebase.

Making a closed-form version that doesn't use `Attribute` lets us
decouple the metaclass hook (which will remain in `attr.rs`) from
the class-field-specific implementation details (which I want to
move) without exposing the intermediate step which is an implementation
detail.

Unfortunately we still need the original helper because
`quantified_instance_as_dunder_call` requires the same processing
but does the lookup in a different way. But the original helper
can now remain private.

Reviewed By: grievejia

Differential Revision: D80626809

fbshipit-source-id: 18bfd0d39b73ebd977af7592a60c798f97333ed5