Fix descriptor inheritance bug
Summary:
We can't look at `.class_object().contains()` to decide about a
descriptor, because that only has the directly-defined class fields
and the descriptor behavior can be inherited.

Fixes #967

It slightly helped the preexisting `_magic_enum_attr` bug, which makes
sense because in 3.11+ that decorator is a descriptor class that
inherits from another descriptor. But something is still not working,
we get `Any` rather than the property-like behavior, so the preexisting
bug is not yet fixed. As far as I can tell we actually are getting the
right answer according to the stubs, so I'll need to investigate more - why
does it work in 3.10 and lower? At any rate, the remaining bug is not related
to this bugfix.

Reviewed By: grievejia

Differential Revision: D80636348

fbshipit-source-id: f051a3ebd57e1a527e820aa21480763ef34317b3