Refactor (TypedDict, TypedDict) check in is_subset_eq in preparation for adding extra_items logic
Summary:
PEP 728 suggests that we treat extra_items as a "pseudo-field" in assignability checks. In preparation for this, move the logic for getting a TypedDict's fields in is_subset_eq into its own helper, and add a TypedDictFieldId enum. The next diff will add an ExtraItems variant to the enum and use it to add the pseudo-field.

For https://github.com/facebook/pyrefly/issues/946.

Reviewed By: stroxler

Differential Revision: D80671109

fbshipit-source-id: 3d7d8a9bccebc6377caf05a56d75dfcf846350f8