Use self to partially instantiate bound methods
Summary:
Context:

A bound method is a callable type that takes a `self` (or `cls`) argument, combined with the `self`/`cls` receiver itself. Whe this function is called, the receiver is applied to the first parameter automatically.

In practice, we can see type parameters in the annotation of this first parameter, and we expect the receiver type to instantiate the type parameter.

Currently, we deal with this during calls by threading the receiver object into the call logic and actually passing it to the parameter, which instantiates type params in the usual way.

However, for subtyping, we just "drop" the first parameter, which means we can't instantiate any type parameters that might exist there.

Instead, we can use the act of binding the method to the receiver, producing a callable, to perform instantiation.

This diff:

When we convert a bound method to a callable, we drop the first parameter, but we also instantiate any tparams present in the first parameter using the bound object.

Reviewed By: stroxler

Differential Revision: D80647786

fbshipit-source-id: 2284f90b037b1b0290c4c4b29ca4ecb456bab52a