Drop first param for unbound generic methods
Summary:
Context:

A bound method is a callable type that takes a `self` (or `cls`) argument, combined with the `self`/`cls` receiver itself. Whe this function is called, the receiver is applied to the first parameter automatically.

In practice, we can see type parameters in the annotation of this first parameter, and we expect the receiver type to instantiate the type parameter.

Currently, we deal with this during calls by threading the receiver object into the call logic and actually passing it to the parameter, which instantiates type params in the usual way.

However, for subtyping, we just "drop" the first parameter, which means we can't instantiate any type parameters that might exist there.

Instead, we can use the act of binding the method to the receiver, producing a callable, to perform instantiation.

This diff:

This diff adds the code to handle generics when dropping the first param of a bound method, but the fix is only partial. If the first parameter of a generic function includes type parameters, we also need to use the bound object to instantiate them.

Reviewed By: stroxler

Differential Revision: D80647789

fbshipit-source-id: 17f605cbf49abf8c2039fd3667f565fd68493a25