Move function binding logic to support generic instantiation
Summary:
Context:

A bound method is a callable type that takes a `self` (or `cls`) argument, combined with the `self`/`cls` receiver itself. Whe this function is called, the receiver is applied to the first parameter automatically.

In practice, we can see type parameters in the annotation of this first parameter, and we expect the receiver type to instantiate the type parameter.

Currently, we deal with this during calls by threading the receiver object into the call logic and actually passing it to the parameter, which instantiates type params in the usual way.

However, for subtyping, we just "drop" the first parameter, which means we can't instantiate any type parameters that might exist there.

Instead, we can use the act of binding the method to the receiver, producing a callable, to perform instantiation.

This diff:

This diff moves the existing "drop first param" logic into the solver, where we can create vars and solve them. In the next diff I will implement the actual instantiation logic.

Reviewed By: stroxler

Differential Revision: D80647788

fbshipit-source-id: 17d136cb42a79882d14d28017bce24e0060870ed