Prepare generic bound method for instantiation with bound object
Summary:
Context:

A bound method is a callable type that takes a `self` (or `cls`) argument, combined with the `self`/`cls` receiver itself. Whe this function is called, the receiver is applied to the first parameter automatically.

In practice, we can see type parameters in the annotation of this first parameter, and we expect the receiver type to instantiate the type parameter.

Currently, we deal with this during calls by threading the receiver object into the call logic and actually passing it to the parameter, which instantiates type params in the usual way.

However, for subtyping, we just "drop" the first parameter, which means we can't instantiate any type parameters that might exist there.

Instead, we can use the act of binding the method to the receiver, producing a callable, to perform instantiation.

This diff:

This diff does two things to prepare for partial instantiation:
1. Replace "tail"-type methods with "split-first" ones, so we have access to the type of the first param
2. Thread the bound object to the right place

At this point, I got stuck, because I need access to the solver to create vars, instantiate them, and substitute. Next diff moves some code around to make this possible. For now, the bound object is passed into the right place, but unused.

Reviewed By: stroxler

Differential Revision: D80647787

fbshipit-source-id: 5f4531f20c8331e19d36a874d3b875c850567be3