Attribute completion / gotodef on modules should look up actual exported names, not `__all__` elements
Summary:
In LSP gotodef / autocompletion, when attribute-access an expression of the form `x.y` where `x` is a module, we currently use the `__all__` elements to decide whether `y` is valid attribute on the module or not.

This is incorrect -- in Python, `__all__` only decides what names to export when downstream file performs a wildcard import (e.g. `from x import *`). In all other cases, `__all__` is ignored and it's the module toplevel static scope that determines what names are exported.

The `__all__` behavior was originally introduced in D70131153 and based on the context I don't think there's any special reason why it has to be `__all__`. Fixing the problem is just a matter of switching to the right API from module exports.

This should fix https://github.com/facebook/pyrefly/issues/865

Reviewed By: rchen152

Differential Revision: D80756773

fbshipit-source-id: b6c75996893016cbb2ce800cce26c27e6f42af1f