Add gt/lt to class field keywords
Summary:
It's unclear to me yet that we can completely get away from checking if a model is a pydantic one downstream of this, but I think that we should collect the values of these keywords and add them to the class field keywords.

Downstream of this, when calculating the field, we can verify their types against the annotation, although like I mentioned, it's not clear that I can get away without verifying if the model is pydantic by adding something more upstream. I will look into that next.

Reviewed By: rchen152

Differential Revision: D80740417

fbshipit-source-id: ad25a484fe7664517f3eb02fdaa18cd7e260e6bc