Validate gt/lt field values against the field annotation
Summary:
This is one of the cases where it seemed reasonable to check if my model is pydantic, but happy to know if there are any alternative thoughts.

The other question is about the error message:
- Does this error message seem accurate?

- Do we need to create pydantic specific error messages, perhaps in a new variant? or should we create general error messages that describe the error and add them to the existing variant?

My thinking is to create general error messages and add them to the existing variant, since those patterns aren't necessarily  exclusive to pydantic, so I would plan to followup with a new error message for this specific error in the next diff. Happy to hear any comments on this.

Reviewed By: rchen152

Differential Revision: D80839388

fbshipit-source-id: bc4ae9317ddf5f37a3e9d01648ad2a90e950a482