Factor enum lookup special case into `enums.rs`
Summary:
It's fine for `attr.rs` to be aware of special-case lookups like enum
`value`, my view is that the `attr` code exists partially to handle
special semantics that aren't part of normal class attribute resolution.

But the details of that lookup are definitely enum-specific, they really
should be in `enum.rs`.

Migrate to a pattern where we have a special case lookup that returns
an optional result, so that we can perform normal class attribute lookups
by intercepting special cases and then, in the typical scenario where we
get no result, using the normal `get_instance_attribute` logic.

This diff preserves existing behaviors, but there are some aspects of
the implementation that we can improve, I'll stack diffs for that.

Reviewed By: grievejia

Differential Revision: D80819999

fbshipit-source-id: 0bfe5834cd8c0bbcb6463107ea7eeb4c0301e9c8