Add a test of mutating `_value_`, which we don't complain about
Summary:
The `value` attribute is actually a property and if it followed normal lookup
rules we would complain; `_value_` really needs to be special-cased.

This is a side-effect of us doing too much in `attr.rs`, where it was tempting
to use `LookupResult.found_type` which allows reads.

We could fix it in place, but I actually noticed it because when I move the logic
out I'm forced to actually construct an `Attribute` at which point this becomes
obvious;

Reviewed By: grievejia

Differential Revision: D80820003

fbshipit-source-id: dda9c44479ab2f37e3ff8b37f7302ccdba250d50