Slightly improve and simplify enum value logic
Summary:
We special-case enum `value` when it isn't annotated, computing the value.

We ought to be doing this for both `_value_` and `value`, and we don't need
to clone `attr_name` to do it because it's the same cost and simpler control
flow to just use a `new_static` in the lookup (which we can eventually optimize
to a static constant, but I want to get the logic out of `attr.rs` first).

The resulting behavior is more correct and the code more ready to be refactored
so that there's no longer an enum-specific hook in `attr.rs`.

Reviewed By: grievejia

Differential Revision: D80812844

fbshipit-source-id: 394adc74b8a025285b4489724b63aca4a7702e13