Remember explicit class targs in bound classmethods
Summary:
Before this diff, we forgot the explicit type argument when creating a bound classmethod, and used a Type::ClassDef instead. When calling the bound method, we hit the `Type::ClassDef(..) <: Type::Type(..)` constraint and instantiate the bound object using `promote_implicit`, producing `Any` (or default).

Instead, if the class has explicit targs, we need to remember them in the bound object type. When calling, the explicit targs can constrain the type parameter in `cls`.

Note that there is still an issue with bound classes without explicit targs. In this case, we still implicitly promote, and still solve type parameters to `Any`. I am documenting this issue, but it remains unsolved for now.

This diff does change the result of `test_self_type_subst_use_receiver`, but I am happy for this change, even if it diverges from mypy and pyright. See this discussion on the Typing discourse forum: https://discuss.python.org/t/calling-methods-on-generic-classes-with-explicit-type-arguments/

Reviewed By: stroxler

Differential Revision: D80829595

fbshipit-source-id: 2bc967c9244552f02d11ac68f29ed4dc8a1f66e9