Rename `TypeVar` attribute cases using `Quantified`
Summary:
As a rule, we use `TypeVar` to mean things that actually
are directly type variables, rather than quantifieds.

Originally, these cases were introduced to support `P.args`
and `P.kwargs`, so the name came from the sense that these
were actually runtime accesses on a type variable directly.

But at this point, the `TypeVar` case is actually just access
on a value of type `Quantified`, so it is best named accordingly,
and the `TypeVarType` case is at least sometimes a value of type
`type[T]` where `T` is a type var.

I opted for `TypeQuantified` rather than `QuantifiedType` because
I consider the former, which evokes `type[Quantified]`, more descriptive;
the latter is ambiguous in that it could be taken to mean some
quantified bounded by `type` which would be a different case.

Reviewed By: yangdanny97

Differential Revision: D80855791

fbshipit-source-id: 51a51b4f62384bd08e8581ba75d2f22ff6905180