Fix empty enum `value` inference
Summary:
It doesn't make sense to have downstream code assume an enum with
no statically-analyzable members is empty - in the presence of dynamic
logic it might not be. It's safer to just use `Any` as the value type.

I realized this might be a general issue - we probably should be careful
assuming `Never` on empty unions, because in cases where the empty union
is due to dynamic logic where we're making strong assumptions, it won't
be very gradual-typing-friendly to assume Never and analyze downstream
that way.

Fixes https://github.com/facebook/pyrefly/issues/984

Reviewed By: yangdanny97

Differential Revision: D80855789

fbshipit-source-id: 29fb02c6dd5e62eb3805d9e14a8cc3e3e4556223