Update some doc comments
Summary:
These were out of date - the original `AttributeBase::TypeVar`
was actually intended to do none of what it does now, it was
originally intended for `P.args` and `P.kwargs`.

But over time, we added logic and eventually migrated that entire
use case to `TypeVarType`, so now none of the doc comments are really
accurate anymore. Patching them up!

Reviewed By: samwgoldman

Differential Revision: D80855794

fbshipit-source-id: a1535260aa31495713df96c7626652635b4dd4d2