Implement completion for unions
Summary:
When the attribute base is a union, we combine suggestions from every branch of the union. This might lead to errors, but it's more helpful to make the suggestion regardless, and let the user deal with the error after, instead of not giving any suggestion.

Removed some stale comments about not expection unions, since that's no longer true now that we have AttributeBase::Union.

Reviewed By: stroxler

Differential Revision: D80887430

fbshipit-source-id: 56b569ecda72c23d0f4e8593957d045a3abde59b