Add override completion test
Summary:
When we walk up the MRO, we keep track of which fields we have seen already, to ensure we don't display the type of overridden fields.

This was already working, but I missed the seen set on my first read, so I wrote the test to confirm. I think the test is worth having, so here we are.

Reviewed By: stroxler

Differential Revision: D80887756

fbshipit-source-id: 5e970d7c2bf6e5fac0daf819222c9027a007b341