Separate AttributeBase1
Summary:
For substituting self during attribute lookup, I want to be able to convert an attribute base to a self type. But, because of AttributeBase::Union, it's awkward.

This diff pulls unions to the top level, into a vec. In the vec, we have "atomic" attribute bases, which are not unions. This gives me the property that I want -- I can get a single attribute base and convert it to a self type.

I appreciate feedback on the naming here, but kindly ask reviewers to be open minded about the base/base1 naming :)

Reviewed By: stroxler

Differential Revision: D80916731

fbshipit-source-id: bc4ac42274c95d0d58d97ab952b821d0c97d973a