add test for bug
Summary:
I meant to write a test case for this when I implemented it but must've forgot.

it's the cause of:
https://github.com/facebook/pyrefly/issues/971

We currently look only in the .py if the module is found. but if the definition exists in .pyi instead of .py, we miss it. since we have already loaded the .pyi, we should look there as a fallback.

Reviewed By: SamChou19815

Differential Revision: D80963222

fbshipit-source-id: c0de8ac18eccbeea02afc6afb7c2658c3e5d17b9