Record "extra" as a dataclass keyword
Summary:
The idea I have is to view dataclasses as having an extra field that's set to false by default.

 I modified `from_type_map` so set `extra` to false by default, which is how dataclasses behave now.

However, I feel like the function name doesn't make sense anymore and that I should followup and rename it. The current name seems to be referring more to the syntax rather than the semantics of the keywords.

We can extract the value from our pydantic model and then mutate the extra dataclass keyword and process it downstream.

My thought there is to add a parameter like so `**kwargs` that can handle any type at `get_dataclass_init`. But I will look into that part more next.

Reviewed By: rchen152

Differential Revision: D80980971

fbshipit-source-id: a90045d46cb61223d06a6f2e21fdb76496d894e2