Reshuffle IDE docs
Summary:
Split the IDE docs into 2 pages as it was getting too long and hard to read. Now we have a separate page for install instructions and supported features, both within an IDE sidebar section (expanded by default).

Also moved the FAQ section to the bottom of the sidebar as its more conventional

Reviewed By: stroxler, connernilsen

Differential Revision: D80820397

fbshipit-source-id: 65de98ba2ecbf2792538b6938b03779019994af8