Add ConfigDict extra unittests
Summary:
Interestingly, with ConfigDict, if the wrong "extra" keyword is provided, the program still runs, unlike in the class keyword case.

I think we should alert the user in both cases. The stubs already do this for us in ConfigDict and I'm not sure we should change that.

There's also the question of whether we should semantically default to extra = true/false in the wrong case, but I feel like it's not that relavant since the program is already problematic (hence to bug tag). I set it to true in the next diff though and will propagate the info to actually remove the error

Reviewed By: stroxler

Differential Revision: D80996984

fbshipit-source-id: 51e18ee57274a60d35dfe3ad4ca0faa9305d5978