Split `Simple` attributes into `ReadOnly` and `ReadWrite`
Summary:
The existing structure is a bit jury-rigged - we actually even forgot to update
the doc comment, which says that `Simple` is always read-write.

We can get rid of the `Visibility` type in the process, which is nice since I want
to introduce another layer to separate out class field semantics; this diff will
help pay for the abstraction cost of adding that layer.

Reviewed By: rchen152

Differential Revision: D80972494

fbshipit-source-id: 13857e1f052b1f728d6bff85764a6f6579fb6fc4