Stop wrapping Type::TypeVar et al. in Type::Type(..)
Summary:
The `T` in `T = TypeVar(..)` is a value, an instance of the `TypeVar` class. Before this diff, we were always wrapping it in `Type::Type(..)`, I assume to make it easier to 'untype' it.

But this choice makes attribute lookup awkward, because it's no longer clear what a `Type::TypeVar` is. Indeed, the code in attr handled both the bare and `Type::Type` wrapped version of this the same way, which doesn't make sense.

Instead, this diff says that `Type::TypeVar` is a value (not Type::Type) but can be used as a type. This resolves some existing issues -- the added test fails before this diff, passes after.

The next diff builds on this to avoid confusion when accessing `P.args` in a type annotation.

Reviewed By: stroxler

Differential Revision: D80992043

fbshipit-source-id: 9adc9ff58f39811223a09a9cf87536774c4fd998