Extract read-write set-and-narrow logic for simple attrs
Summary:
I'm working on trying to split the logic for Attribute vs ClassAttribute.

The idea is that Attribute handles all kinds of wierd language-level semantics
like gradual type attribute lookups, module lookups, `__getatttr__` and
`__setattr__` fallback, distributing over unions, etc. This is a lot, and it's
nice to aggregate it all in one layer of the code.

Class attributes, on the other hand, have to deal with another large but in
principle disjoint set of complexity around descriptors, all of the semantics
(many of them open to debate) around instance vs class access, and so on.

One of the harder operations to split out is set-and-narrow... the fact that
we can potentially narrow to a different type for every element of a union base
due to contextual typing makes this pretty complicated even though most real
examples are simple.

Here, I am extracting the core logic - which for now is only used on read-write
attributes - into a helper because once I move the ReadOnly and ReadWrite cases
to ClassAttribute, we'll need to share some code here.

The new helper is responsible for the actual contextual subtyping of a simple
attribute set and applying the narrow.

Reviewed By: grievejia

Differential Revision: D81011735

fbshipit-source-id: 2538ec64236fc0c12b20dec178d791af246b7102