Introduce a `ClassAttribute` datatype
Summary:
This diff introduces a `ClassAttribute` datatype that I eventually want to
use to split class attribute semantics from `Attribute` - this will enable
`class_field` logic to operate with fewer layers of abstraction and
transformation.

Unfortunately unless I do one giant commit that changes everything, things
have to get arguably messier before they become cleaner because a lot of
functionality is currently defined in terms of `Attribute` that eventually
needs to move to `ClassAttribute` but it's hard to do that in one change.

This diff is trying to introduce the new datatype with minimal pain, and so in
order to avoid duplication and hard-to-read match statements, I'm delaying
dealing with "simple" read-only and read-write class attributes by just leaving
those as top-level `Attribute`... I need to extract some helpers (which I'll do
next diff) before it makes sense to deal with these.

At this stage it's just a mechanical transformation where we migrated the
variants that *only* make sense for class fields to a new type, no meaningful
structural changes have happened yet.

Reviewed By: grievejia

Differential Revision: D81011733

fbshipit-source-id: f649e1292a0bfa4c6d6c8d68f978e56e918ffb52