Add ReadOnly / ReadWrite to ClassAttribute
Summary:
This diff finishes building out the ClassAttribute type algebra
by moving the ReadOnly and ReadWrite cases from AttributeInner to
ClassAttribute.

We still do need a simple "read write type" for a number of existing
uses that are not normal class attributes, for example
- attribute access on Any and Never
- module attributes
so I introduce `AttributeInner::Simple` for that; all existing callsites
were read-only (whether this is correct is debatable, but I'm currently
focused on the refactor so we can defer that question) so we don't currently
need a read-write case that isn't a class attribute.

Most of the logic refators quite nicely. The subset check is pretty ugly,
but it doesn't matter long-term because soon we'll be able to make attribute
lookups on classes use `ClassAttribute`, at which case the type signatures
will change to something we can structure much nicer.

Reviewed By: grievejia

Differential Revision: D81011736

fbshipit-source-id: 24c76c84c309845a7f6fb9f0be5ab9a8f704132e