Factor out class-attribute-delete helper. Improve comments
Summary:
Mypy allows `del` on just about anything.

Given our current focus on adoption, and the fact that `del` is pretty rare in real
code except for power users with memory-intensive programs anyway, it makes sense
to follow that.

This diff doesn't change any functionality, but it breaks class_attr logic into a
helper and updates the comments based on the observation that mypy is pretty lax
about `del`, eliminating a TODO for now (I think we can let this be unless we
get a bug report, it doesn't seem like something worth being proactively pedantic
about).

Reviewed By: grievejia

Differential Revision: D81011737

fbshipit-source-id: 472254f05e1d8c2da5777f483a36de58619c2018