Extract `is_class_attribute_subset` logic
Summary:
Most of the attribute subset logic is really only relevant for comparing
two class attributes, which makes sense since the main question is whether
some class is a subtype of protocol.

We do have to have the ability for `want` to be another type, for example
`Any`, `Never`, or a `Module`, but in those cases we either hardcode that
the subtype check fails, or we treat the attributes we find "as if" they were
instance lookups, by synthesizing a ClassAttribute.

The resulting code is modular in that we can treat ClassAttribute as a lower
layer in the logic (my ultimate goal is to move this to `class_field.rs` and be
able to change the representation as needed), and the match is also just easier
to read as two matches rather than a complex match using nested data.

Reviewed By: grievejia

Differential Revision: D81012666

fbshipit-source-id: 5f3d4b99be96fd84416c971a121138d9d55f33aa