Rework smart constructors: add found_class_attribute, remove found_read_only
Summary:
If we want to properly separate the class attribute and attribute layers, having
smart constructors that try to cross the boundary is messy.

It's cleaner instead to directly pass the class attribute, and in the process
add a `found_class_attribute` helper to LookupResult that actually saves us far
more boilerplate than `found_read_only` ever did.

Reviewed By: grievejia

Differential Revision: D81044556

fbshipit-source-id: 7d58d02862319afe07c369e31a6bd82c0a1afb74