Add some plumbing to cut class_field code over
Summary:
At this stage, ClassAttribute expresses all the cases that I wanted to extract
out of Attribute. The next stage will be to cut over various functions in
`class_field` to produce values of type `ClassAttribute` instead of on
`Attribute`.

That's going to be a big diff, getting some trivial plumbing out of the way first
makes it easier: we're going to need all the same smart constructors we have
today for `Attribute`, so that the `class_field` logic can use them. And we'll
need the `ClassAttribute` datatype to be public.

Make both those changes in this diff.

Reviewed By: grievejia

Differential Revision: D81011738

fbshipit-source-id: f8b8cabc6534b3cefe46c786a5a6084e3d95923d