Simplify is_protocol_subset_at_attr
Summary:
The protocol subset logic is pretty silly given the current state of
things - we know already that every name we look up is a protocol
member, which means we're doing a plain instance lookup on it, but
we thread that through multiple layers of abstraction to get an
`Attribute` instead of a `ClassAttribute`, which requires:
  - an otherwise unnecessary clone of the protocol ClassType, just
    so that we can thread an owned AttributeBase through plumbing
  - the existence of `try_lookup_attr_from_class_type`, whose name
    alone is a hint that something smelly is going on
  - in the process of all this, we just forget about a bunch of
    restrictions on the returned type, which then forces downstream
    code to deal with a bunch of impossibilities like getting a
    non-ClassAttribute result.

Let's start cleaning this up. This diff leaves the Attribute based
logic in place for now, but makes it much clearer that this isn't
necessary by just using an instance lookup and wrapping in
`Attribute::class_attribute`, instead of threading the logic through
`try_lookup_attr_from_class_type`.

In the process, we're able to delete `try_lookup_attr_from_class_type`
and eliminate a clone.

Reviewed By: grievejia

Differential Revision: D81012320

fbshipit-source-id: 615c5fcb7feae8822907f7e4e8824679e53b8d2b